firmware_update: resolve technical debt
The goal is to decouple and concentrate generic logic from the two
firmware update mechanisms currently present in the code-base. This
should make it more easy for new devs to append new transport and
therefore support more devices.

This commit also updates the rockusb dependency to the fork containing
fixes so that the RK1 devices work. fixes #45,
https://github.com/turing-machines/BMC-Firmware/issues/45
