+++ tpi_rs/Cargo.toml
@@ -15,8 +15,8 @@ evdev = { version = "0.12.1", features = ["tokio"] }
 futures = "0.*"
 gpiod = { version = "0.2.3", default-features = false }
 once_cell = "1.18.0"
-rockfile = { version = "0.1.0", optional = true } 
-rockusb = { version = "0.1.1", optional = true }
+rockfile = { version = "0.1.0", optional = true }
+rockusb = { git = "https://github.com/svenrademakers/rockchiprs.git", branch="main", optional = true }
 rusb = "0.9.2"
 rustpiboot = { git = "https://github.com/ruslashev/rustpiboot.git", rev="89e6497", optional = true }
 serde = { version = "1.0.180", features = ["derive"] }
@@ -25,6 +25,7 @@ sqlx = { version = "0.7.1", features = [
     "sqlite"
 ] }
 futures-concurrency = "7.4.1"
+
 anyhow.workspace = true
 log.workspace = true
 simple_logger.workspace = true
