+++ tpi_rs/src/middleware/firmware_update/mod.rs
@@ -1,27 +1,27 @@
 #[cfg(feature = "rock")]
-mod rk1_fwudate;
+mod rockusb_fwudate;
+#[cfg(feature = "rock")]
+use self::rockusb_fwudate::new_rockusb_transport;
 #[cfg(feature = "rpi")]
 mod rpi_fwupdate;
+#[cfg(feature = "rpi")]
+use rpi_fwupdate::new_rpi_transport;
+pub mod transport;
+use self::transport::FwUpdateTransport;
+
 use super::usbboot::FlashProgress;
 pub use super::usbboot::FlashingError;
 use futures::future::BoxFuture;
 use once_cell::sync::Lazy;
-#[cfg(feature = "rock")]
-use rk1_fwudate::Rk1FwUpdateDriver;
-#[cfg(feature = "rpi")]
-use rpi_fwupdate::RpiFwUpdate;
 use rusb::GlobalContext;
 use std::collections::HashMap;
-use tokio::{
-    io::{AsyncRead, AsyncSeek, AsyncWrite},
-    sync::mpsc::Sender,
-};
+use tokio::sync::mpsc::Sender;
 
 #[allow(clippy::vec_init_then_push)]
 pub static SUPPORTED_MSD_DEVICES: Lazy<Vec<(u16, u16)>> = Lazy::new(|| {
     let mut supported = Vec::<(u16, u16)>::new();
     #[cfg(feature = "rpi")]
-    supported.push(RpiFwUpdate::VID_PID);
+    supported.push(rpi_fwupdate::VID_PID);
     supported
 });
 
@@ -29,36 +29,29 @@ pub static SUPPORTED_DEVICES: Lazy<HashMap<(u16, u16), FactoryItemCreator>> = La
     let mut creators = HashMap::<(u16, u16), FactoryItemCreator>::new();
     #[cfg(feature = "rpi")]
     creators.insert(
-        RpiFwUpdate::VID_PID,
+        rpi_fwupdate::VID_PID,
         Box::new(|_, logging| {
-            Box::pin(async move {
-                RpiFwUpdate::new(logging)
-                    .await
-                    .map(|u| Box::new(u) as Box<dyn FwUpdate>)
-            })
+            Box::pin(async move { new_rpi_transport(logging).await.map(Into::into) })
         }),
     );
 
     #[cfg(feature = "rock")]
     creators.insert(
-        Rk1FwUpdateDriver::VID_PID,
+        rockusb_fwudate::RK3588_VID_PID,
         Box::new(|device, logging| {
             let clone = device.clone();
-            Box::pin(async {
-                Rk1FwUpdateDriver::new(clone, logging).map(|u| Box::new(u) as Box<dyn FwUpdate>)
-            })
+            Box::pin(async { new_rockusb_transport(clone, logging).await.map(Into::into) })
         }),
     );
 
     creators
 });
 
-pub trait FwUpdate: AsyncRead + AsyncWrite + AsyncSeek + std::marker::Unpin + Send {}
-pub type FactoryItem = BoxFuture<'static, Result<Box<dyn FwUpdate>, FlashingError>>;
+pub type FactoryItem = BoxFuture<'static, Result<Box<dyn FwUpdateTransport>, FlashingError>>;
 pub type FactoryItemCreator =
     Box<dyn Fn(&rusb::Device<GlobalContext>, Sender<FlashProgress>) -> FactoryItem + Send + Sync>;
 
-pub fn fw_update_factory(
+pub fn fw_update_transport(
     device: &rusb::Device<GlobalContext>,
     logging: Sender<FlashProgress>,
 ) -> anyhow::Result<FactoryItem> {
